refactor: split formatter utils
