+++ packages/store/src/index.ts
@@ -16,7 +16,6 @@ export * from './space.js';
 export * from './store.js';
 export * from './text-adapter.js';
 export type { IdGenerator } from './utils/id-generator.js';
-import type * as Y from 'yjs';
 export type { Y };
 export {
   createAutoIncrementIdGenerator,
@@ -27,6 +26,10 @@ export {
 export * as Utils from './utils/utils.js';
 export * from './workspace/index.js';
 export * from '@blocksuite/global/utils';
+
+import './utils/formatter.js';
+
+import type * as Y from 'yjs';
 
 const env =
   typeof globalThis !== 'undefined'
